Concat with grunt-ts and fix HTMLInputElement errs
Dropped concat for putting together the JavaScript files, because
the same function can be achieved with grunt-ts. This means that
errors can be narrowed down to individual files, rather than the
full concatenated TS file. Also cast a few HTMLElements as
HTMLInputElements to take care of a few compiler warnings. There
are still a few more errors to work through, so it's not building
correctly just yet.