Resets photo index on opening new place infowindow
If you opened a place, and scrolled through to, say, its 5th photo, then opened a second place that only had two photos, then the 'Prev/Next' links would throw an error. Now the currentPhoto property is reset to 0 whenever a new place infowindow is opened, which fixes this bug.