Moves directionsDisplay to DirectionsPanel
This property was originally a global variable, and has been hanging around main.ts for a long time. It was being passed into a chain of functions whose only purpose seemed to be getting it into some of the DirectionsPanel methods.

Now, it's simply a static member of DirectionsPanel. This helps remove it from main.ts, and simplifies the signatures of several related and unrelated methods.